@@ -26,9 +26,9 @@
  * example, to map https://www.drupal.org:8080/mysite/test to the configuration
  * directory sites/example.com, the array should be defined as:
  * @code
- * $sites = array(
+ * $sites = [
  *   '8080.www.drupal.org.mysite.test' => 'example.com',
- * );
+ * ];
  * @endcode
  * The URL, https://www.drupal.org:8080/mysite/test/, could be a symbolic link
  * or an Apache Alias directive that points to the Drupal root containing
